7#ERRORH transient error fixed

+++ src/main/resources/db/migration/V6__resources.sql
@@ -1,5 +1,5 @@
 CREATE TABLE resources (id INT AUTO_INCREMENT PRIMARY KEY,
                         amount SMALLINT (10),
                         type VARCHAR(45),
-                        updated_at SMALLINT (10),
+                        updated_at BIGINT (12),
                         kingdom VARCHAR (45));
